% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwts2ph.generator.R
\name{nwts2ph.generate}
\alias{nwts2ph.generate}
\title{This file genreate the example dataset nwts2ph
importFrom("stats", "rbinom")}
\usage{
nwts2ph.generate(data, seed = 20)
}
\arguments{
\item{data}{the full cohort data}

\item{seed}{the random seed we use for generating this dataset}
}
\description{
This file genreate the example dataset nwts2ph
importFrom("stats", "rbinom")
}

